+++ CHANGELOG.md
@@ -273,4 +273,4 @@ ajv                        6.9.1   →   6.10.0
 ## [1.0.0] 2018-01-29
 ### Original Release
 - Added Material-UI as base framework
-- Added design from Material Dashboard by Creative Tim
+- Added design from Material Dashboard